Refactor so both dirty code modals use the same component for presentation (but not logic)
